Remove WalkthroughActivityLabelTest and add the test case in WalkthroughActivityTest
