+++ app/src/main/AndroidManifest.xml
@@ -215,7 +215,7 @@
       android:theme="@style/OppiaThemeWithoutActionBar" />
     <activity
       android:name=".app.walkthrough.WalkthroughActivity"
-      android:label="@string/walk_through_activity_title"
+      android:label="@string/walkthrough_activity_title"
       android:theme="@style/OppiaThemeWithoutActionBar" />
     <activity android:name=".app.testing.LessonThumbnailImageViewTestActivity" />
     <activity
