+++ app/src/main/res/values/strings.xml
@@ -380,7 +380,7 @@
   <!-- CompletedStoryListActivity -->
   <string name="completed_story_list_activity_title">Stories Completed</string>
   <!-- WalkthroughActivity -->
-  <string name="walk_through_activity_title">Walk Through</string>
+  <string name="walkthrough_activity_title">App walkthrough</string>
   <string name="walkthrough_welcome_description">Learn new math skills with stories that show you how to use them in your daily life</string>
   <string name="welcome">"Welcome %s!"</string>
   <!-- WalkthroughTopicListFragment -->
